Add files via upload
Implementation of Predicting Human Activity Using Convolutional Neural Network
The dataset used is the data taken from the accelerometer of a smartphone. The rest of the details are mentioned in the readme file of the dataset. The link to the dataset is given in the code above.

The attached file is a Jupyter notebook and I took help from the mentioned Youtube video: https://www.youtube.com/watch?v=lUI6VMj43PE&list=LL&index=8
Some of the methods were deprecated and so I made the changes and now the project is up-to-date.
Happy Learning, Thanks. xx